+++ src/parser.ts
@@ -1,5 +1,5 @@
 import Parsimmon, { takeWhile, alt, index, lazy, makeSuccess, notFollowedBy, of, Parser, regex, seq, seqMap, seqObj, string, whitespace, any } from 'parsimmon'
-import { basename } from 'path'
+import { basename, dirname } from 'path'
 import unraw from 'unraw'
 import { Assignment as AssignmentNode, Body as BodyNode, Catch as CatchNode, Class as ClassNode, Describe as DescribeNode, Entity as EntityNode, Expression as ExpressionNode, Field as FieldNode, If as IfNode, Import as ImportNode, List, Literal as LiteralNode, Method as MethodNode, Mixin as MixinNode, Name, NamedArgument as NamedArgumentNode, New as NewNode, Node, Package as PackageNode, Parameter as ParameterNode, Program as ProgramNode, Reference as ReferenceNode, Return as ReturnNode, Self as SelfNode, Send as SendNode, Sentence as SentenceNode, Singleton as SingletonNode, Super as SuperNode, Test as TestNode, Throw as ThrowNode, Try as TryNode, Variable as VariableNode, Problem, Source, Closure, ParameterizedType as ParameterizedTypeNode } from './model'
 import { mapObject, discriminate } from './extensions'
@@ -101,8 +101,14 @@ export const File = (fileName: string): Parser<PackageNode> => {
       imports: Import.sepBy(optional(_)).skip(optional(_)),
       members: Entity.sepBy(optional(_)),
     }).skip(optional(_))
-  )
+  ).map(filePackage => {
+    const dir = dirname(fileName)
+    return (dir === '.' ? [] : dir.split('/')).reduceRight((entity, name) =>
+      new PackageNode({ name, members:[entity] })
+    , filePackage)
+  })
 }
+
 
 export const Import: Parser<ImportNode> = node(ImportNode)(() =>
   key('import').then(obj({